Add info about MySQL config ONLY_FULL_GROUP_BY
Added information in the fasp-api installation to explain how to change
MySQL config to allow GROUP BY queries without an alias. This needs to
be done in the mysql config file but in the Docker container it can
only currently be set temporarily as the standard MySQL docker container
doesn't have any config files.